Revise JNDI Context wiring
Introduced DynamicJndiContextFactoryBuilder to create a JndiContextFactory which can handle dynamic JNDI lookups from multiple sources.
If the lookup environment is empty, we return a JndiContextFactory that returns a InitialContext which supports a lookups against a fixed set of Spring beans.
If the environment is not empty, we try to use the provided java.naming.factory.initial classname to create a JndiContextFactory and
delegate further lookups to this instance.

This also fixes the LDAP User Federation problems from #11.

Fixes #11