Revise Keycloak starter configuration
We now ship with an embedded default keycloak configuration
to ease the use of the starter. Users can customize the
keycloak configuration via the application configuration file as usual.

We now generate a random admin password at startup if an admin
user should be created but no password was set.

Fixes #24